# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cycle of Kings game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /towns/{townId}:
    get:
      operationId: getTown
      tags: [towns]
      summary: Get town details including resources
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Town
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Town"
        "404":
          description: Not found
  /towns/{townId}/grid:
    get:
      operationId: getTownGrid
      tags: [towns]
      summary: Get the 9x9 building grid for a town
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Grid cells
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GridCell"
  /towns/{townId}/buildings:
    post:
      operationId: placeBuilding
      tags: [towns]
      summary: Place a new building on the grid
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingPlacement"
      responses:
        "201":
          description: Building placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GridCell"
        "400":
          description: Invalid placement or insufficient resources
  /towns/{townId}/buildings/{row}/{col}:
    patch:
      operationId: moveBuilding
      tags: [towns]
      summary: Move building to a new position
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
        - name: row
          in: path
          required: true
          schema: { type: integer }
        - name: col
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuildingMove"
      responses:
        "200":
          description: Building moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GridCell"
        "400":
          description: Invalid move
    delete:
      operationId: removeBuilding
      tags: [towns]
      summary: Remove/refund a building (75% resource refund)
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
        - name: row
          in: path
          required: true
          schema: { type: integer }
        - name: col
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Building removed, refund granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Town"
        "404":
          description: No building at that position
  /towns/{townId}/buildings/{row}/{col}/upgrade:
    post:
      operationId: upgradeBuilding
      tags: [towns]
      summary: Start an upgrade on a building
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
        - name: row
          in: path
          required: true
          schema: { type: integer }
        - name: col
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Upgrade started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GridCell"
        "400":
          description: Cannot upgrade (insufficient resources, already upgrading, max level)
  /towns/{townId}/fortifications:
    get:
      operationId: getFortifications
      tags: [towns]
      summary: Get all fortifications for a town
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Fortifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fortification"
    post:
      operationId: placeFortification
      tags: [towns]
      summary: Place a wall segment or tower
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FortificationPlacement"
      responses:
        "201":
          description: Fortification placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fortification"
        "400":
          description: Invalid placement
  /towns/{townId}/fortifications/{row}/{col}:
    delete:
      operationId: removeFortification
      tags: [towns]
      summary: Remove a fortification
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
        - name: row
          in: path
          required: true
          schema: { type: integer }
        - name: col
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Town"
...
  /towns/{townId}/trades:
    get:
      operationId: getTownTrades
      tags: [trade]
      summary: Get active trade routes for a town
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Trade routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeRoute"
    post:
      operationId: createTradeRoute
      tags: [trade]
      summary: Create a trade route with another town
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeRouteInput"
      responses:
        "201":
          description: Trade route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeRoute"
        "400":
          description: Invalid trade route
  /towns/{townId}/trades/{tradeId}:
    delete:
      operationId: cancelTradeRoute
      tags: [trade]
      summary: Cancel a trade route
      parameters:
        - name: townId
          in: path
          required: true
          schema: { type: integer }
        - name: tradeId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled
...
components:
  schemas:
...
    Town:
      type: object
      required: [id, playerId, name, gold, food, wood, stone, goldPerHour, foodPerHour, woodPerHour, stonePerHour, defenseRating, population, populationCap, lastTickAt]
      properties:
        id: { type: integer }
        playerId: { type: integer }
        name: { type: string }
        gold: { type: number }
        food: { type: number }
        wood: { type: number }
        stone: { type: number }
        goldPerHour: { type: number }
        foodPerHour: { type: number }
        woodPerHour: { type: number }
        stonePerHour: { type: number }
        defenseRating: { type: number }
        population: { type: integer }
        populationCap: { type: integer }
        lastTickAt: { type: string }
...
    GridCell:
      type: object
      required: [id, townId, row, col, buildingType, level, upgrading, upgradeEndsAt]
      properties:
        id: { type: integer }
        townId: { type: integer }
        row: { type: integer }
        col: { type: integer }
        buildingType:
          type: string
          enum: [farm, mine, quarry, barracks, market, tavern, house, empty]
        level: { type: integer }
        upgrading: { type: boolean }
        upgradeEndsAt: { type: ["string", "null"] }
    BuildingPlacement:
      type: object
      required: [row, col, buildingType]
      properties:
        row: { type: integer }
        col: { type: integer }
        buildingType:
          type: string
          enum: [farm, mine, quarry, barracks, market, tavern, house]
    BuildingMove:
      type: object
      required: [newRow, newCol]
      properties:
        newRow: { type: integer }
        newCol: { type: integer }
    Fortification:
      type: object
      required: [id, townId, row, col, type, level]
      properties:
        id: { type: integer }
        townId: { type: integer }
        row: { type: integer }
        col: { type: integer }
        type:
          type: string
          enum: [wall, tower]
        level: { type: integer }
        borderBonus: { type: boolean }
    FortificationPlacement:
      type: object
      required: [row, col, type]
      properties:
        row: { type: integer }
        col: { type: integer }
        type:
          type: string
          enum: [wall, tower]
...
    TradeRoute:
      type: object
      required: [id, fromTownId, toTownId, fromTownName, toTownName, resourceType, amountPerHour, active, createdAt]
      properties:
        id: { type: integer }
        fromTownId: { type: integer }
        toTownId: { type: integer }
        fromTownName: { type: string }
        toTownName: { type: string }
        resourceType:
          type: string
          enum: [gold, food, wood, stone]
        amountPerHour: { type: number }
        active: { type: boolean }
        createdAt: { type: string }
    TradeRouteInput:
      type: object
      required: [toTownId, resourceType, amountPerHour]
      properties:
        toTownId: { type: integer }
        resourceType:
          type: string
          enum: [gold, food, wood, stone]
        amountPerHour: { type: number }
...