# lib/api-zod/src/generated/types/gridCell.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cycle of Kings game API
 * OpenAPI spec version: 0.1.0
 */
import type { GridCellBuildingType } from './gridCellBuildingType';

export interface GridCell {
  id: number;
  townId: number;
  row: number;
  col: number;
  buildingType: GridCellBuildingType;
  level: number;
  upgrading: boolean;
  /** @nullable */
  upgradeEndsAt: string | null;
}