# lib/api-zod/src/generated/types/tradeRoute.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cycle of Kings game API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRouteResourceType } from './tradeRouteResourceType';

export interface TradeRoute {
  id: number;
  fromTownId: number;
  toTownId: number;
  fromTownName: string;
  toTownName: string;
  resourceType: TradeRouteResourceType;
  amountPerHour: number;
  active: boolean;
  createdAt: string;
}