# lib/api-zod/src/generated/types/tradeRouteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cycle of Kings game API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRouteInputResourceType } from './tradeRouteInputResourceType';

export interface TradeRouteInput {
  toTownId: number;
  resourceType: TradeRouteInputResourceType;
  amountPerHour: number;
}